refactor: implement expandProducts utility to enhance product filtering by size and color
